==> real data analysis/00_dependencies.R <==
library(dplyr)
library(fansi)
library(nlme)
library(tidyverse)
library(Deriv)
library(stringr)
library(LaplacesDemon)
library(purrr)
library(MASS)
library(mvtnorm)
library(tibble)
library(ggpubr)
library(ecoreg)
library(foreach)
library(parallel)
library(doParallel)
library(Matrix)
library(here)
library(batchtools)
library(R.utils)
